chore: scope auth-logic in seperated module
